Remove unused utilities from SharedTree
