+++ packages/dds/tree/src/id-compressor/index.ts
@@ -56,3 +56,4 @@ export {
     assertIsUuidString,
 } from "./uuidUtilities";
 export { SessionIdNormalizer } from "./sessionIdNormalizer";
+export { fail, Mutable, getOrCreate, compareStrings } from "./utils";
