+++ packages/dds/tree/src/id-compressor/utils.ts
@@ -73,6 +73,22 @@ export function compareMaps<K, V>(
 }
 
 /**
+ * Retrieve a value from a map with the given key, or create a new entry if the key is not in the map.
+ * @param map - The map to query/update
+ * @param key - The key to lookup in the map
+ * @param defaultValue - a function which returns a default value. This is called and used to set an initial value for the given key in the map if none exists
+ * @returns either the existing value for the given key, or the newly-created value (the result of `defaultValue`)
+ */
+export function getOrCreate<K, V>(map: Map<K, V>, key: K, defaultValue: (key: K) => V): V {
+    let value = map.get(key);
+    if (value === undefined) {
+        value = defaultValue(key);
+        map.set(key, value);
+    }
+    return value;
+}
+
+/**
  * Compares strings lexically to form a strict partial ordering.
  */
 export function compareStrings<T extends string>(a: T, b: T): number {
