+++ packages/dds/tree/src/test/id-compressor/idCompressor.perf.spec.ts
@@ -23,8 +23,9 @@ import {
     LocalCompressedId,
     OpSpaceCompressedId,
     SessionId,
+    fail,
+    Mutable,
 } from "../../id-compressor";
-import { fail, Mutable } from "../../util";
 import {
     Client,
     IdCompressorTestNetwork,
