+++ packages/dds/tree/src/util/index.ts
@@ -34,24 +34,15 @@ export { StackyIterator } from "./stackyIterator";
 export {
     clone,
     compareArrays,
-    compareFiniteNumbers,
-    compareFiniteNumbersReversed,
-    compareMaps,
     compareSets,
-    compareStrings,
-    copyPropertyIfDefined,
     fail,
     getOrAddEmptyToMap,
-    getOrCreate,
-    hasAtLeastLength,
     isJsonObject,
     isReadonlyArray,
     JsonCompatible,
     JsonCompatibleObject,
     JsonCompatibleReadOnly,
     makeArray,
-    Mutable,
     RecursiveReadonly,
-    setPropertyIfDefined,
     unreachableCase,
 } from "./utils";
